1.2.3 - Item syncing and Handling
Added a bunch of Item related data structures:
ItemStackMap: an item type map which stores all items in a HashMap based on type, essentially a roster of every item stored.
ItemSlotMap: an item type map which stores lists of slots containing specific items. Optimizes item extraction code significantly.
FastItemStackHandler: a version of the forge IItemHandler implementation which uses the 2 maps above for optimization.

SyncableInventory and SyncableItemHandler:
Inventory is a SyncableData that implements IItemHandler.
Handler is a SyncableValue that contains a FastItemStackHandler.