1.2.5 Significant Rework of tons of stuff.
SyncableData:
- Removed unnecessary separation between the SyncableData and SyncableValue classes. SyncableData is now SyncableCompound and extends SyncableValue.
- SyncableData classes of all types have been removed and replaced with respective SyncableValues.

TileEntities:
- Removed a bunch of base VCTEs. The syncable system is all that is necessary for Tile entity differentiation.
- Added PhantomTEs and PhysicalTEs. These will be the only ViciousCore base entities.

PhysicalTEs: in terms of how tile entities normally work, physical tiles are nothing special.
PhantomTEs: Phantom Tile entities are special in that they use a PhantomMemory object instead of a SyncableCompound to store data. (PhantomMemory is a SyncableCompound wrapper) using a PhantomMemory object is important to PhantomTEs as they are kept track of in the PhantomMemoryManager and persist even after tile destruction. This means that when used properly, tiles can continue functioning even after the chunk has unloaded. Of course though, this isn't recommended for all tiles as depending on how much time your tile's code can take this can lower server TPS. Using PhantomMemory also allows for using the same data in different tiles. A good example of this is with Tesseracts, 2 tesseracts could use the same data when connected to the same channel meaning changing the state of one will also change the state of the other.

PhantomMemory:
PhantomMemory is a way of storing data constantly outside of the game's data storage.
PhantomMemoryTickable is a form of the above with a tick function.